use AniList as image assets for `/anime`
## Considerations

AniList uses no-logo poster/cover image on their database, which is
suitable for thumbnail preview on Discord (to avoid artifacting
if seen afar).

Background provided from AniList is not using conventional 16:9/16:10
ratio (which both SIMKL and Kitsu used), but uses 1024px width and
up to 500px height, which is looking better in embed view.

## Fallback Method

If AniList can not find an entry, use SIMKL as fallback, then
continued using Kitsu.

Howeverm if all three sites does not have an entry of a title
currently looked up, final fallback to MyAnimeList